+++ .gitignore
@@ -151,3 +151,11 @@ cython_debug/
 #  option (not recommended) you can uncomment the following to ignore the entire idea folder.
 #.idea/
 .vscode/launch.json
+codegen/codegenlaunch.json
+test/core/output_compare/*
+test/core/GraphQLClients/RapidApi/downloaderArgs.json
+test/core/GraphQLClients/GithubApi/generatorArgs.json
+test/core/GraphQLClients/RapidApi/generatorArgs.json
+test/core/GraphQLClients/GithubApi/downloaderArgs.json
+test/core/GraphQLClients/GdbcApi/generatorArgs.json
+test/core/GraphQLClients/GdbcApi/downloaderArgs.json
