--- test/GraphQLClients/GdbcApi/Model/countries.py
@@ -1,8 +1,0 @@
-from pygqlmap import GQLConnection, GQLEdges
-from .country import countryNode
-
-class countries(GQLConnection):
-    
-    def __init__(self):
-        super().__init__(GQLEdges(countryNode())) 
-
