--- test/GraphQLClients/GdbcApi/Model/countryRegions.py
@@ -1,7 +1,0 @@
-from pygqlmap import GQLConnection, GQLEdges
-
-class countryRegions(GQLConnection):
-    
-    def __init__(self):
-        super().__init__(GQLEdges()) 
-
