--- test/GraphQLClients/GdbcApi/Model/populatedPlaces.py
@@ -1,8 +1,0 @@
-from pygqlmap import GQLObject
-
-class populatedPlaces(GQLObject):
-    totalCount: int
-    
-    def __init__(self):
-        self.totalCount = 0
-        super().__init__()
