+++ test/core/GraphQLClients/GdbcApi/Model/countryRegions.py
@@ -1,0 +1,7 @@
+from pygqlmap.components import GQLConnection, GQLEdges
+
+class countryRegions(GQLConnection):
+    
+    def __init__(self):
+        super().__init__(GQLEdges()) 
+
