+++ test/core/GraphQLClients/GdbcApi/Model/currencies.py
@@ -1,0 +1,10 @@
+
+from GraphQLClients.GdbcApi.Model.currency import currency
+from pygqlmap.components import GQLConnection, GQLEdges
+
+
+class currencies(GQLConnection): 
+    
+    def __init__(self):
+        super().__init__(GQLEdges(currency())) 
+
