+++ test/core/GraphQLClients/GdbcApi/Model/currency.py
@@ -1,0 +1,14 @@
+
+from pygqlmap.components import GQLObject
+
+
+class currency(GQLObject): 
+    countryCodes: list
+    code: str
+    symbol: str
+    
+    def __init__(self):
+        self.countryCodes = []
+        self.code = ''
+        self.symbol = ''
+        super().__init__()
