+++ test/core/GraphQLClients/GdbcApi/Model/locales.py
@@ -1,0 +1,9 @@
+# from sgqlc.types import Field, list_of
+# from .common import GDBCConnection, GDBCEdge
+# from .locale import Locale
+
+# class LocaleEdge(GDBCEdge):
+#     node: Field(Locale)
+
+# class LocalesConnection(GDBCConnection):
+#     edges: list_of(LocaleEdge)
