+++ test/core/GraphQLClients/GdbcApi/Model/timeZone.py
@@ -1,0 +1,9 @@
+# from sgqlc.types import String, Int, ID, Type, Field, non_null
+# from sgqlc.types.datetime import DateTime, Time
+# from sgqlc.types.relay import Node
+
+# class TimeZone(Node):
+#     name: Field(non_null(String), graphql_name='name')
+#     rawUtcOffsetHours: Field(non_null(Int), graphql_name='rawUtcOffsetHours')
+#     dateTime:Field(non_null(DateTime), graphql_name='dateTime')
+#     time:Field(non_null(Time), graphql_name='time')
