+++ test/core/GraphQLClients/GdbcApi/Model/timeZones.py
@@ -1,0 +1,9 @@
+# from sgqlc.types import Field, list_of, non_null
+# from .common import GDBCConnection, GDBCEdge
+# from .timeZone import TimeZone
+
+# class TimeZonesEdge(GDBCEdge):
+#     node = Field(non_null(TimeZone), graphql_name='node')
+
+# class TimeZonesConnection(GDBCConnection):
+#     edges = Field(non_null(list_of(TimeZonesEdge)), graphql_name='edges')
