+++ test/core/consts.py
@@ -1,0 +1,19 @@
+gdbcHeaders = {
+  "content-type": "application/json", ##it needs to be added for graphQl version
+	"X-RapidAPI-Key": "Api Key",
+	"X-RapidAPI-Host": "geodb-cities-graphql.p.rapidapi.com"
+}
+
+gdbcUrl = "https://geodb-cities-graphql.p.rapidapi.com/"
+
+githubHeaders = {"Authorization": 'Api Key'}
+
+githubUrl = "https://api.github.com/graphql"
+
+rapidApiHeaders = {
+        "content-type": "application/json", ##it needs to be added for graphQl version
+	    "X-RapidAPI-Key": "Api Key",
+	    "X-RapidAPI-Host": "graphql-rapidapi-test.p.rapidapi.com"
+        }
+
+rapidApiUrl = 'https://graphql-rapidapi-test.p.rapidapi.com/'
