+++ test/core/requirements.txt
