+++ test/helpUnitTest.py
@@ -1,0 +1,11 @@
+
+import os
+
+async def runGeneratorCommandHelp():
+    print('\nRunning runGeneratorCommandHelp...')
+    command = "CodeGenerator -h" #command to be executed
+    print("Launching: " + command)
+    
+    ret = os.system(command)
+    print("End of runGeneratorCommandHelp")
+
